requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langgraph
langchain-groq
langchain-huggingface
langchain-community

faiss-cpu

tavily-python
pydantic
PyPDF2

faiss-cpu
sentence-transformers